Enable devMode with react-native init-windows old/* templates
