+++ packages/@react-native-windows/cli/src/generator-windows/index.ts
@@ -203,6 +203,8 @@ export async function copyProjectTemplateAndReplace(
     packageGuid: packageGuid,
     currentUser: currentUser,
 
+    devMode: options.useDevMode,
+
     useExperimentalNuget: options.experimentalNuGetDependency,
     nuGetTestFeed: options.nuGetTestFeed,
     nuGetADOFeed: nugetVersion.startsWith('0.0.0-'),
@@ -230,12 +232,7 @@ export async function copyProjectTemplateAndReplace(
       ? [
           // app common mappings
           {
-            from: path.join(
-              srcRootPath,
-              options.useDevMode
-                ? 'metro.devMode.config.js'
-                : 'metro.config.js',
-            ),
+            from: path.join(srcRootPath, 'metro.config.js'),
             to: 'metro.config.js',
           },
           {
