--- vnext/template/metro.devMode.config.js
@@ -1,56 +1,0 @@
-const {getDefaultConfig, mergeConfig} = require('@react-native/metro-config');
-
-const fs = require('fs');
-const path = require('path');
-const exclusionList = require('metro-config/src/defaults/exclusionList');
-
-const rnwPath = fs.realpathSync(
-  path.resolve(require.resolve('react-native-windows/package.json'), '..'),
-);
-
-// [devMode
-const rnwRootNodeModules = path.resolve(rnwPath, '..', 'node_modules');
-const rnwPackages = path.resolve(rnwPath, '..', 'packages');
-// devMode]
-
-/**
- * Metro configuration
- * https://facebook.github.io/metro/docs/configuration
- *
- * @type {import('metro-config').MetroConfig}
- */
-
-const config = {
-  // [devMode
-  watchFolders: [rnwPath, rnwRootNodeModules, rnwPackages],
-  // devMode]
-  resolver: {
-    blockList: exclusionList([
-      // This stops "react-native run-windows" from causing the metro server to crash if its already running
-      new RegExp(
-        `${path.resolve(__dirname, 'windows').replace(/[/\\]/g, '/')}.*`,
-      ),
-      // This prevents "react-native run-windows" from hitting: EBUSY: resource busy or locked, open msbuild.ProjectImports.zip or other files produced by msbuild
-      new RegExp(`${rnwPath}/build/.*`),
-      new RegExp(`${rnwPath}/target/.*`),
-      /.*\.ProjectImports\.zip/,
-    ]),
-    // [devMode
-    extraNodeModules: {
-      'react-native-windows': rnwPath,
-    },
-    // devMode]
-  },
-  transformer: {
-    getTransformOptions: async () => ({
-      transform: {
-        experimentalImportSupport: false,
-        inlineRequires: true,
-      },
-    }),
-    // This fixes the 'missing-asset-registry-path` error (see https://github.com/microsoft/react-native-windows/issues/11437)
-    assetRegistryPath: 'react-native/Libraries/Image/AssetRegistry',
-  },
-};
-
-module.exports = mergeConfig(getDefaultConfig(__dirname), config);
